semitransparent countdown overlay PLUS - fixed the old inter-round timeouts in Timer.js for when the Done isn't clicked - separated out the logic to calculate and show the round results
