+++ public/assets/js/components/Timer.js
@@ -1,11 +1,13 @@
 import Life from '../components/Life.js';
 import Rounds from '../components/Rounds.js';
+import Inventory from '../components/Inventory.js';
 
 import calculateGameResults from "../helpers/calculateGameResults.js";
 import roundCountdown from "../helpers/roundCountdown.js";
 import clearBoardForNewRound from "../helpers/clearBoardForNewRound.js";
 import restingMode from "../helpers/restingMode.js";
 import Players from "../helpers/Players.js";
+import tallyMoves from "../helpers/tallyMoves.js";
 
 let myTimer = document.querySelector('div.timer-counter');
 let myTimerCounter = document.querySelector('span.time-nums');
@@ -15,7 +17,7 @@ class Timer {
 
     // timePerRound;
 
-    counterRange = [0, 30];
+    counterRange = [0, 10];
 
     element;
 
@@ -45,53 +47,29 @@ class Timer {
             this.counter--;
             // this.element.innerHTML = this.counter;
             if (this.counter === 0) {
-                const opponentMove = Players.all.player2.generateRandomMoves();
-                console.log(opponentMove)
+                tallyMoves();
 
-                let opponentTallyColumn = document.querySelectorAll(`table.tally.opponent-tally td:nth-child(${Rounds.all['game1'].counter})`);
-                console.log('opponent', opponentTallyColumn);
-                opponentTallyColumn.forEach((td, index) => {
-                    let moveComponent = Object.values(opponentMove)[index];
-                    if (moveComponent === 'attack') {
-                        td.classList.add('cell-attacked');
-                    } else if (moveComponent === 'block') {
-                        td.classList.add('cell-blocked');
-                    }
-                });
-
-                let myTallyColumn = document.querySelectorAll(`table.tally.my-tally td:nth-child(${Rounds.all['game1'].counter})`);
-
-                if (Object.values(opponentMove).includes("attack")) {
-                    myTallyColumn.forEach(td => {
-                        td.classList.add('round-defeat');
-                    });
-                    opponentTallyColumn.forEach(td => {
-                        td.classList.add('round-won');
-                    });
-                    Life.all.myLife.decreaseCounter();
-                } else {
-                    myTallyColumn.forEach(td => {
-                        td.classList.add('round-draw');
-                    });
-                    opponentTallyColumn.forEach(td => {
-                        td.classList.add('round-draw');
-                    });
-                }
-
-                if (Rounds.all['game1'].counter < Rounds.all['game1'].counterRange[1] && Life.all.myLife.counter > 0 && Life.all.opponentLife.counter > 0) {
+                let leftPlayerTotalInventory = Inventory.all['attack-left'].counter + Inventory.all['block-left'].counter;
+                let rightPlayerTotalInventory = Inventory.all.opponentAttack.counter + Inventory.all.opponentBlock.counter;
+                if (
+                    Rounds.all['game1'].counter < Rounds.all['game1'].counterRange[1] 
+                    && Life.all.myLife.counter > 0 
+                    && Life.all.opponentLife.counter > 0
+                    && leftPlayerTotalInventory + rightPlayerTotalInventory !== 0
+                ) {
                     restingMode();
-                    this.resetCounter();
-
-                    setTimeout(() => {
+                    setTimeout(async () => {
                         document.querySelector("div.countdown-overlay").classList.remove("d-none");
-                        roundCountdown();
-                    }, 5000)
-
-                    setTimeout(() => {
+                        await roundCountdown();
+                        // Timer.all['myTimer'].startCounter();
                         document.querySelector("div.countdown-overlay").classList.add("d-none");
                         Rounds.all['game1'].increaseCounter();
-                        clearBoardForNewRound(Rounds.all['game1'].counter);
-                    }, 9000);
+                        // console.log(Rounds.all['game1'].counter)
+                        setTimeout(() => {
+                            clearBoardForNewRound(Rounds.all['game1'].counter);
+                        }, 800);
+                    }, 1600);
+                    this.resetCounter();
                 } else {
                     restingMode();
                     this.resetCounter();
